Added lang and delete, fixed issues, improved ui

+++ app/Http/Controllers/ReviewController.php
@@ -26,14 +26,14 @@ class ReviewController extends Controller
     public function store(ReviewRequest $request) : RedirectResponse
     {
         $user = Auth::user();
+        /** @var User $reviewedUser */
         $reviewedUser = User::find($request->user_id);
         $existingReview = $reviewedUser->reviews->firstWhere('reviewer_id', $user->id);
 
-        if ($existingReview) {
-            return back()->withErrors([
-                'rating' => 'You have already reviewed this user',
-            ]);
+        if ($existingReview || $user->id == $reviewedUser->id) {
+            return back();
         }
+
         if ($request->validated()) {
             $review = new Review();
             $review->fill($request->validated());
@@ -66,9 +66,18 @@ class ReviewController extends Controller
         }
         return redirect()->route('reviews.index', $review->user_id);
     }
-    public function destroy(string $id): void
+    public function destroy(string $id): RedirectResponse
     {
         $review = Review::find($id);
+        if (auth()->id() != $review->reviewer_id) {
+            return redirect()->back();
+        }
         $review->delete();
+        $reviewsLeft = Review::where('user_id', $review->user_id)->count();
+        // If there are no reviews left, redirect to profile view
+        if ($reviewsLeft == 0) {
+            return redirect()->route('profile', ['url' => $review->user->url]);
+        }
+        return redirect()->back();
     }
 }
